requirements.txt
einops==0.6.1
fastclasses_json==0.6.1
matplotlib==3.6.3
numpy==1.24.1
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
Requests==2.31.0
scikit_learn==1.2.0
seaborn==0.12.2
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.64.1
transformers==4.25.1
